Layout UI test improvements
This CL improves the unit tests for layout rendering a bit (there are
now simpler ways to assert the resource folder for the current layout,
and a simpler way to track subsequent render updates). It also adds a
base LayoutTest test project where we can check in more complicated
layouts for use by layout tests, and finally this CL adds a layout
rendering test which exercises the configuration switcher a bit more,
checking that the preferred device list is working correctly.

Change-Id: I78bbd70e36d4bf7c8d8c61fdfb42a7fdee96d1f0