SDK merging from SDK switch dialog on project sync
When Android Studio notices a mismatch between the project SDK and
its main SDK during project sync, it currently displays a dialog
asking the user which one to switch everything to. This CL extends
that by examining both SDKs, and if the chosen SDK is missing
packages extant in the other one (or its versions are out-of-date),
it puts up a dialog offering to copy the bits over into the chosen
SDK. It does this copy via a cancelable background task.

Change-Id: I6cabe9e0759abe0b7369f6d9a7062fe80a1c1b85