Install the best split apk for device if split info is available
Gradle model 0.13 introduces the concepts of split apks where an
apk is specific only to a particular combination of density or abi.
This CL adapts Studio to first look at whether the current gradle
model supports splits, and if so, use that information to identify
the best apk to install.

In a future CL will introduce a bigger reorganization to only build
the specific split apks that are necessary.

Change-Id: I051a6bf6c69108e885943b1e605c5b686c677cfa