Change ApkUploader to not use SyncService
ApkUploader class used SyncService to upload an apk. However,
at the end of uploading, the service was not closed, resulting
in a descriptor leak. This can be observed by doing:
  - start Studio
  - install and run apk
  - disconnect device (so that all client connections are closed)
  - lsof | grep 5037, and notice that a connection is in CLOSE_WAIT

All that needed to be done was to call service.close() at the end.
This CL changes the ApkUploader service to call device.pushFile()
rather than expliclty obtaining the sync service as we weren't
really making use of the sync service's progress monitor callbacks.

Possible fix for b/14562106 and
https://code.google.com/p/android/issues/detail?id=74651

Change-Id: Idecacbe3a71f1655fbe9553a6f9bf943f5583fb9