Add more diagnostics to help track down index corruption bug
This adds toString methods for a value class that appears in the
error message for this cache corruption bug:
76569: Index query code crashes with "Expected only one value per-inputId"

Knowing more about the kinds of resource entries that are involved
might help us ask follow-up questions to figure out what kinds of
data can trigger this.

Also optimize the code which recreates a set with a known size: don't
use the HashSet contructor with a fixed size, since due to the default load
factor adding that number of elements will cause a map resize. Similarly,
size the identifier StringBuilder capacity before building it.

Change-Id: Ic6cdf65d5a998ae144752e162113dfeb8c64644b