Use dumpsys package output while determining if an app is already installed
ApkUploader service is used by the launch code to determine if
an app needs to be re-uploaded and installed on the device. It worked
by maintaining an internal state of what apks were installed on what
device. The existing logic had two issues as identified by
https://code.google.com/p/android/issues/detail?id=76539:

 1. It assumed that an apk was installed as soon as it was uploaded
    to the device. However, uploading and installing were two separate
    operations, and it was possible for the launch to be stopped in
    between the two operations.
 2. It has no knowledge of any operations on the device outside of
    Studio. (e.g. uninstalling an app outside Studio).

This CL fixes both the issues: The API has now been changed to provide
two methods: isInstalled and setInstalled. It is now the callee's
responsibility to upload and install based on the result of isInstalled,
and update the state by calling setInstalled after a successful install.

The second issue is fixed by invoking "dumpsys package <id>" on the
device and making sure that the details match from the previous installation
via Studio. This implies that the cache hit path now has the additional
overhead of querying the package manager, but that seems to be less
than a hundred milliseconds.

Change-Id: I5b8e00e385dfa7c550b27c306ea2cee8e608d88f