Assign unique id's in the layout editor
Until now, the layout editor has managed id's in layouts on a per
layout basis, meaning that the first button dropped will be assigned
"button1", the next one "button2", and so on -- even if there is a
second layout which also contains these id's.

That's generally safe, but has been a massive source of confusion
because if you *rename* an id, that obviously needs to rename the same
id globally (to update the associated activity findViewById calls and
so on). There have been a number of bugs filed complaining about id's
changed in one layout affecting id's in other files.

While we can't solve this problem in general (if you have a layout
which references "submit_button", and you try to change this id to
"post_button", we'll need to update code references across the app; we
can't figure out a way to "only" update the references that apply to
this layout), we can mitigate the problem somewhat: we now assign
unique id's when you drop new widgets. At least that way, if you drop
a button, which just happens to have the same id as an existing
button, your subsequent attempt to rename the button won't affect any
other layouts.

This cleans up the id handling in the layout editor a bit, since we no
longer need to store a persistent list of id's-per-layout (and the
list would not always be accurately kept up to date anyway).

Also adds workaround for a rare crash, and adds an explicit
orientation to LinearLayouts created as part of Tabs and SidingDrawers
(since without these you get a lint warning.)

Change-Id: I1920fa9ce210aa7c0a0c6f33fe7d0db4f892a217