Fix for: Switching from phone/tablet resets the tablet layout
https://code.google.com/p/android/issues/detail?id=76912

'Unattached' activities are activities that are neither the source nor
target of any transition.

Unattached activities are stacked over the top of each other because,
in the case of apps like ioshed, the general graph layout algorithm
creates a layout that is far too wide to be usable.

That said, if the user positions such activities manually, those positions
should be persisted and respected on reload as with ordinary activities.

This CL fixes that problem. 

Change-Id: Idd49c246cf8c43544ae7cad04a7cee7c79a37780