Refactor: Move the package containing the Navigation Editor's 'public API' into the IDE
With this CL we cross the point of no return on Navigation Editor architecture:
Navigation Editor will be code generation-based and not have an external
library, external file format or public API.

Move all of the classes that were intended to become part of a public API back
under the IDE package structure. So:

com.android.navigation -> com.android.tools.idea.editors.navigation.model

Change the Navigation Model reader to ignore the errors in reading old
navigation files (thereby discarding the previously stored x/y co-ordinates)
and start from a computed layout.

This is being done now as the file structure for supporting multi-module
projects was different from the single-module system anyway: doing this
at the same time means we trash layouts once rather than twice. Not worth
spending time on converters at this stage but there's a place for them for
when there are more users.

Change-Id: Id43a3b4f0827e2ba0a6ba2687aed69241b8e0144