Fixed infinite sync with empty gradle.settings file.
CLs https://android-review.googlesource.com/107590 and
https://android-review.googlesource.com/108422 added a
check for empty settings.gradle where if found, the IDE
will warn the users and ask them if they'd like to
continue with project sync. This was a workaround for
https://code.google.com/p/android/issues/detail?id=72294
and back then I thought this was a bug in Gradle.

After doing some debugging, it turns out that the bug is
in Android Studio, where the listener that gets notified
when a sync is finished was not finishing project setup
when an empty settings.gradle file was present.

In this CL:
- Moved code that sets up projects during sync from
  GradleProjectImporter to a new class, ProjectSetupTask.
  This makes the code easier to maintain.
- When we detect that the project is non-Android Gradle
  project, we finish project setup (this is the fix
  for the bug.)
- Removed code in PreSyncChecks that checked if a
  settings.gradle file is empty (by parsing the contents
  using the Groovy lexer) and checked if a project
  is a multi-module project (by counting the number
  of build.gradle files in the project.) This code
  was quite fragile.

Fixes
https://code.google.com/p/android/issues/detail?id=72294
finally!

Change-Id: I8b75051ff8e37bb04eb2cb1d55971d775e7ebb27
(cherry picked from commit 64d7af9)