Fix style DOM inspector
Updates the quick documentation for the style resolver, which needed
some fixes after the recent StyleResourceValue changes (in particular,
getNames() on that method now includes android: as a prefix for framework
names, so iterating on getNames() as keys didn't work. Use the new getItems
to iterate over ItemResourceValues instead. Also fixes a formatting issue
(some items were not separated by newlines), and since we have the info,
include android: as a prefix for framework attribute names.

Change-Id: I2159b8aa5cd4f91eadd9a8d50b1bb301d21c5ca5