Bug 76727: Use bigger icons in new model gallery
This change will now use 128x128 icons in the new module type gallery.
I am now trying to access all form factor icons through
DeviceMenuAction#FormFactor with a long term goal of moving it to
FormFactorUtils.

There was also a bug in the ASGallery when cell margins were not set.
This change required fixing it.

Change-Id: I12116d28989b2a2dbb00f48b4b3b5461957c9ae0