73030: Avoid dropping a TableRow on an existing row.
Currently, dropping a TableRow in an existing row will wrap the new one
into the existing TableRow. LayoutLib will generate an NPE when this
happens.
This fix workarounds that by avoiding nesting TableRows.

Also fixing a problem with the coordinate system used in
calculateGridInfo. This was causing the TableLayout editor to not work
correctly when there was a zoom other than 1.
Now calculateGridInfo operates only the the Component target coordinate
system.
Added some comments trying to clarify when the model coordinates are
used.

Change-Id: Ic0b6d7b8b4165599a6745fad1c52b6d9d53f85a4