Fix HAXM detection
The call ExecUtil.execAndGetOutput takes a command line
as a list; we were passing in a single element in that
list which contained space separated arguments. That doesn't
work; the first argument needs to be just the command.

Also add in version detection on OSX such that users on
for example 1.0.8 will be notified that 1.1.1 is available.

Change-Id: I0830e3c8f349fe1d948993c022e8e7739e285047