Ensure app namespace is imported with custom GridLayout
If you have a v7 GridLayout that does not import the
app namespace, and you drop a child widget into it, the
layout editor will incorrectly set the layout_row and
layout_column attributes. This CL imports the auto-res
namespace if necessary such that they are used for the
layout attributes.

(Related to issue 78352)

Change-Id: Ib1d450dd0ca014b3dbfdd411b4710b26b9a77190