Do not close emulator when closing Studio
A previous CL introduced the feature of showing a console with
the output from the emulator. This console is shown in the Run
tool window, and behaves just like it is connected to a running
process (the emulator). If a user closes the IDE when an external
process is running, the IDE now shows a dialog asking whether
it should detach from the process, and whether that process
should be terminated.

However, the emulator process cannot be killed since it will
leave stale lock files around. We also want to retain the previous
behavior that a running emulator should be independent of the
IDE. So this CL makes the IDE silently detach from the emulator
process and leave it running even when the IDE is closed.

Fixes https://code.google.com/p/android/issues/detail?id=78401

Change-Id: Ie50a8c0b928cc8b37706061596150b195610cd53