Fixed bug where sync will add wrong dependencies to build.gradle files.
The fix is actually reverting the functionality added by this CL
https://android-review.googlesource.com/108763 . Instead of reverting
the CL, I'm just unregistering the project root listener that
triggers the change in build.gradle files. This way we keep the
infrastructure that Denis added. Denis will apply a proper fix later.

This CL fixes blocker bug
https://code.google.com/p/android/issues/detail?id=78628
and reopens
https://code.google.com/p/android/issues/detail?id=73938

I'll submit a UI test for this later today, to make sure we don't
reintroduce this bug in the future.

Change-Id: Ied846994b4b8807367abaf56f7db9665deb60117