Now we don't notify that local.properties is updated when setting SDK.
In this CL we no longer notify that the local.properties file is going
to be updated with the path of the newly selected Android SDK (from the
"Project Structure" dialog.) The reason is that this behavior is
expected. We still notify that the local.properties of other open
projects (i.e. not the current project) are going to be modified
(the SDK path is a global setting.)

Fixes
https://code.google.com/p/android/issues/detail?id=65643

Change-Id: I89ec733db799e7704dcfd0114b65db34fb4d2f1c