Work around invalid index sizes
This CL works around a corrupt index bug I've encountered once or twice
(it just happened again), and it's also been reported by users. This fix
explicitly throws an IOException (rather than relying on the RuntimeException
thrown from the HashMap call). That exception type is deliberately called
by the indexing machinery and causes the index to be rebuilt, which is what
we want in this scenario.

We check not just for negative values, but for large positive values too,
since with a corrupt index you may be unlucky and get a large positive
random value instead (which can trigger the OutOfMemoryException dialog.)

I ran into this for the AndroidIdIndex, but I've also seen reports for the
AndroidValueIndex which has the same issue.

java.lang.IllegalArgumentException: Illegal initial capacity: -638446103
     at java.util.HashMap.<init>(HashMap.java:172)
     at java.util.HashMap.<init>(HashMap.java:199)
     at java.util.HashSet.<init>(HashSet.java:125)
     at com.intellij.util.containers.HashSet.<init>(HashSet.java:33)
     at org.jetbrains.android.AndroidIdIndex$2.read(AndroidIdIndex.java:94)
     at org.jetbrains.android.AndroidIdIndex$2.read(AndroidIdIndex.java:82)

Change-Id: Idf5e6918bd4a7a9e63195e61a7b4ebce615cca73