67348: WrongViewCast sometimes generates incorrect warnings
The wrong view cast detector would sometimes generate invalid
warnings that a view cast was wrong. This happened when a view
id is used with more than one type of widget in different layout
files, and the lint check is running incrementally within the IDE.

For incremental analysis, the lint check relies on the resource
repository to look up all the occurrences of a resource item,
and a bug in the resource repository merger (which combines
resources from different source sets and makes for example resource
items defined in a debug source set hide same-named resources in
the default config) would cause only a single resource id definition
to survive the merge. That meant the lint check would see only a
single view type for a given id and would conclude that other view
casts were wrong.

Change-Id: I9273160fe78eca63c9895fa38a2673f14d11c8b7