apk installation: do not throw error for unexpected dumpsys output
After installing an apk on a device, we perform a "dumpsys package"
on the installed package to obtain the installation timestamp. This
was used to identify whether an apk needs to be installed again in
the future.

Unexpected output from the dumpsys command would result in an error
being logged. This CL changes the severity of that log to a warning.

Fix for https://code.google.com/p/android/issues/detail?id=79778

Change-Id: Ia4575a9872ce0b2ca4126df5e8b0004ef2604c29
(cherry picked from commit c301c8d9e03e45cccdcc160341654e7fd2bda3a6)