81768: Compute default parameter values quickly
Old code for computing default parameter values was meant to be
invoked lazily, but this never worked due to Free Marker specifics.
Sideeffect was that defaults computation was invoked recursively and
seems to have ridiculuos complexity (it looked like exponential growth
depending on number of template parameters - debugging shown up to
140000 computations in one invocation).
Proposed code has worst-case n^2 (were n is a number of parameters),
with the common case being O(n) because in our templates default
parameter values depend on static values and can all be computed in
one pass.

Change-Id: I256d4822d19f5001e422b3c9bc29b12d925e7c56