Updating SD card control in AVD manager and fixing layout
Previously the SD card control in the AVD edit dialog was unusual and
confusing. Updating to be clearer.
Also changed the concept slightly to "built-in" vs. "external" cards,
where the "built-in" one is the image in the avd folder, and
"external" means specified separately.
Also cleaning up the layout of that page. Unfortunately it seems that
intellij doesn't deal with col-span > 1 for gridlayout panels that
inherit the grid from a parent panel, so flattening that dialog out.
Fixes https://code.google.com/p/android/issues/detail?id=82764
and https://code.google.com/p/android/issues/detail?id=82991

Change-Id: I79969d470ea84af6768fdb21e2692088b5b82262