Installer now waits for uninstaller
The NSIS uninstaller, by default, extracts itself to a temporary
directory and runs in the background, so waiting on its execution
to complete doesn't do anything.

This CL changes the call to the uninstaller to execute itself
directly, meaning the installer now actually waits for the
uninstall to finish.

Related issue:
  https://code.google.com/p/android/issues/detail?id=82203

Change-Id: Id0ebd1a1e8015e3f3d24674a8254b456e8bab4a0