Enable reopening of translations editor at startup
IJ reopens files at startup if they were open while closing. The
translations editor wasn't reopened in this manner because of two
reasons: The virtual file backing the editor could not be found, and the
editor was not DumbAware. This CL fixes this issue by:

 - Implementing a new AndroidFakeFileSystem registered with the IDE.
   This file system knows how to recreate a StringsVirtualFile on
   demand.
 - The StringsResourceEditor is marked DumbAware

Fix for https://code.google.com/p/android/issues/detail?id=75472

Change-Id: I9c6f93c167db872500f0df93303f4b7f70140329