Template merging: Reduce creation of temporary projects
GradleFileMerger relies on PSI to merge Gradle files, which requires
a valid project. If it is not passed in a valid project, then it creates
a temporary project (named "MergingOnly"), and attempts to perform
operations with that project. It turns out that in newer versions of
IJ (140.x branch), this immediately causes a IndexNotReady exception.

This CL attempts to avoid this scenario by passing in a valid project
reference whenever possible. This eliminates the main issue with 140.x,
allowing the creation of projects. There are still other exceptions of
the same form thrown later during template rendering (such as when
using the formatter during template execution, which relies on the Java
code formatter).

Change-Id: I8b56152fb5d2a67d11cf6aeb36cb00a92ef1ffde
(cherry picked from commit 01c28f5)