Pick test runner from Gradle while creating test configuration
While running tests, we use the testInstrumentationRunner specified
in Gradle if nothing is specified in the run configuration. This is
slightly confusing because when you look at the run configuration
itself, the test runner field is empty even though the instrumentation
runner from Gradle is used.

This CL improves this behavior by initializing the run configuration
with the instrumentation runner from Gradle when the run configuration
is created.

Fix for https://code.google.com/p/android/issues/detail?id=98997

Change-Id: I4f42ae7ff427d685cef6f3122bd75d08a7ead77d