Adds Capture tool window to show all profiling captures
A secondary tool window next to the project view can now
be used to show all the captures taken from a device.

When this tool window is opened, it monitors changes to
the 'captures' directory inside the project folder.
All our profiling tools will be migrated to drop captures
files in this directory.

Change-Id: I3f1175aa9ed989d7edebe791c99d084ca8070242