Recognise only layouts whose top-level tag is "Space" as empty
Previously layouts whose top-level tag ended in "Layout" and which had no
children were recognised as empty layouts whose activities would not be
shown at all in the Navigation Editor. This is too permissive as many of
the existing templates in the 'NewTemplateWizard' create layout files
that are FrameLayouts with no children - where the children are added
in code. Activities using this idiom should not be excluded from the
overview of the application flow provided by the Navigation Editor.
Instead then, a layout file must use the tag "Space" at the top level
to be excluded from the Navigation View.

Change-Id: Ieacf69554465c525c96b68bafefc1de0ceab593c