Factored out state list traversal out of resolveMultipleX functions
Previously, both resolveMultipleDrawables and resolveMultipleColors
implemented breadth-first traversal of resource references and references
in state lists, and it was mostly duplicated code, specialized in a couple
of places.

This CL introduces StateListIterable, which implements independent
breadth-first traversal of state lists & resource references, and changes
both resolveMultipleDrawables and resolveMultipleColors to use it.

If this approach looks good, it could be extended even further,
parameterizing by traversal strategy (a boolean flag which would indicate
whether we want all options from state list or only default / latest one),
which will allow to re-use it also in resolveColor and resolveDrawable.

Change-Id: If4d087765ae82218c9bfc627fa81de995fe49962