Fix PSI to Lombok conversion for enum constants with initializers
The conversion would skip converting the initializer from the PSI
tree to the Lombok tree. This surfaced as a bug in the unused resource
detector which failed to look at resources referenced only in an
enum constant initializer.

The code to do the migration was already there. I think it must have
been commented out by accident; there was no unit test covering that
scenario. There is now.

Change-Id: Iaec9f3e5040cb0c1b2307bc3991061751871ec8c