wizard: Fix test that is broken on Jenkins.
This test mysteriously did not fail locally, but failed on Jenkins 
because an Application instance was created which caused a different
code path of execution.
This change uses AndroidTestBase as the test base class instead of 
TestCase. An Application instance is created to mimic the results seen 
on Jenkins.
The bug fix is to test for isUnitTestMode() and NOT require to be 
running on UI thread.

Change-Id: Ie06f3414b5abbdfcec37df8f2ced0182a6e1d8d9
(cherry picked from commit 31b4714a740a8997a15ad331ce6a292439bf4bc0)