IDE can now find and include a module containing a selected file.
When using the "Project subset" feature (it is currently
experimental and disabled by default") it is possible that
the user does not know the module she wants to include,
but knows the file she wants to work on. It would be handy
to be able to include the module that contains a selected
file, by just right-clicking it.

Fixes
https://code.google.com/p/android/issues/detail?id=161000

Change-Id: I0f5c82d21a765e28ed5166f918671988d0aa9896