Introduce CodeTemplate class to store templates instead of using PsiMethod
To simplify creation of templates (which are used for detection of transitions
and code generation), IntelliJ's parsing capabilities are used: templates
are represented as methods, where parameter list correspond to template
parameters and method body should contain a single statement that consists of
an expression, which corresponds to template body.

Prior to this CL, raw PsiMethod instances were passed as templates, and
extraction of template body (an expression from the only statement) happened
at use point.

This CL introduces special class for storing code templates, which could be
created from PsiMethod instance to use existing templates with little
modification.

Change-Id: Id34220e31bf88349dea75051aa80c3c8ab6474fa