Now we correctly set dependencies on aars exposed as artifacts.
The issue was that, when a project exposes an aar file as an artifact,
we were treating the aar as a jar file, and just creating a library
pointing directly to the aar, instead of the jar and res folder in
the exploded aar. The paths of the jar and res folder were in the model
but we were ignoring them, and setting the module dependency directly.

The fix involves:
1. Not creating libraries for aar files exposed as artifacts.
2. If module A depends on module B, but module B exposes aar libraries
   as artifacts, we now link to the jar and res folder inside the
   aar file (the paths of these files are in AndroidProject.)

Fixes
https://code.google.com/p/android/issues/detail?id=162634

Change-Id: Iadb905d52d05571fbf0c7a7863f04dc67456f771