Unable to create a new Android Library
We were setting the form factor for the Android Library template
to null, but then later trying to use it to determine what wizard
path would be used. Setting it back to Mobile.
This has the unfortunate side effect of raising the priority of
Android Library in the gallery. I will check in this change to unblock
the creation of libraries, and then file a bug for the incorrect
ordering.

Change-Id: I7c05e526dc9571975364015fad84d4d7bad45ba3