Developer Services and Project Structure
This CL provides all the core classes and structure to support
developer services, a concept of enabling APIs and associated utility
code within a projet. The UI for developer services is shown inside
the Project Structure dialog, and the services themselves are defined
by a service.xml file and provided by plugins.

Change-Id: Iee8136e958283c9714db7412bbca2654869364a3