IDE now detects if Android plugin versions are incompatible with Gradle versions.
Android plugin 1.0.+ and 1.1.+ are not compatible with Gradle 2.4.
This CL verifies that if Gradle 2.4 is being used, the IDE notifies the user
if the project is using an incompatible version of the Android plugin, and
offers a quick fix that changes the plugin version to 1.2.

The version of Gradle used by the project is obtained this way:
1. From one of the Gradle models, at sync time (this is functionality
   added by this CL)
2. If #1 fails, the IDE tries to get the Gradle version from the
   project configuration (wrapper or local distribution)

Change-Id: If156e1759483a55041ceeafccae7a697f712fa0e