Fix R$id class creation for aars.
When the project is not compiled, the generated R$id class only had ids
for file resources and value resources. Any ids generated via "@+id" in
other resources were not created. This made it impossible to preview a
new project which used appcompat.

Change the generation of the class to use the R.txt file for id
generation. For all other resource types, still parse the contents of
the exploded aar.

Change-Id: Ic7684597db0b3879b7acc1eefb8cc3273799a16e