Fix AndroidColorAnnotator.colorToString
Previously, AndroidColorAnnotator.colorToString have used
Long.toString, which doesn't show trailing zeroes. That lead to
weird-looking results when alpha value is less than 16.
Now, alpha chanel would be shown always with two digits.

Also, this CL adds a couple of tests to this function, makes this
function public (it would be useful in couple of other places) and
moves it to ResourceHelper.

Change-Id: Id47c3be36f2b04e29f8a94dada27233aeec28890