Now IDE validates path of local Gradle distribution to use.
If the path to the local Gradle distribution is not set, or if
it does not belong to a directory, the IDE will ask the user
to use the Gradle wrapper instead (the preferred way to use
Gradle.) If the user does not want to use the wrapper, the IDE
will let the user choose the path of a local Gradle
distribution.

This CL also removes unused code in PreSyncChecks class.

Change-Id: Ibc6d3fd71b821d010730839f831d1c29e375da74