InferNullable: Override the Infer action with Android action.
In Android Studio we want to update the gradle build file instead
of making Java library changes.
This change goes with the change: 149844 in the tools/idea project
for using the Android annotations by default.
This class would be simpler if InferNullableAction from Intellij
allowed more reuse. We should make a request for intellij to change
this action such that it is easier to override.

(cherry picked from commit cd3eed24f522a32fadd1686e4dc82465feaba9fb)

Change-Id: I869ee637a5b25b7af3a789cd30636f71b9ae490e