Remove the dedicated AndroidSdkLangaugeLevelPusher.
Currently, android sdk sources are configured in a way that the ide
builds stubs eagerly for them (caused by
AndroidSdkLanguageLevelPushe.updateSdkLanguageLevel()), which result
in the problem that the same class in android SDK would be found
twice in many places.

In comparison, the Java sdk won't push langugage level to its source
file which does not suffers this problem, so that's behavior we want
for android SDK also.

Note that don't build stubs eager doesn't mean the stub won't be
built, when you go to a class file, IDEA could find the corresponding
java file and build the stub, in this way, the duplication problem
won't happen.

Fix:
https://code.google.com/p/android/issues/detail?id=78011

Change-Id: Iffa6cd19edcb6d160e427ad2857c6bb3f8318c4b