run config: restore preferred avd correctly
The run config editor has a combo box that displays a list of AVDs.
An earlier CL changed this combobox to use IdDisplay[] as the model
item rather than String[]. However, the save/restore functionality
depended on the item being a string. As a result, restoring previous
state was not working properly.

This CL fixes the issue by comparing the avd id string to the id as
obtained from IdDisplay instead of comparing them directly. It also
sets the renderer once during creation as opposed to doing it
conditionally.

Fixes https://code.google.com/p/android/issues/detail?id=170450

Change-Id: I2120826c2bcaaeed01d854bee93d675111178180