Nele: Add speed search to the palette.
Add speed search and make the tree the receiver of the focus when
the palette window is chosen. The latter allows speed search to
be started after the user clicks on the titlebar of the palette.

Change-Id: Ifbf2a815ffcebf939fbe1a1fb0286bf834434a98
(cherry picked from commit 54df1dc070c983d459e327050f096a5fd02d28a7)