Removed double Gradle sync when creating new modules.
When creating new modules (Android library or Java library) 2
gradle syncs get triggered: one from Template and another one
from NewFormFactorModulePath.

Since the one from Template is not just used for module
creation but for other files as well (e.g. a new Activity) we
cannot remove the call to Gradle sync from there. We can
not remove the call to Gradle sync from the wizards, because
the open files after a new module is created.

The fix is to add a new overloaded version of "render" in
Template that takes a flag that indicates whether Gradle sync
should be executed if needed. The "New Module" wizards set
this flag to false, because they take care of Gradle sync
themselves.

Fixes
https://code.google.com/p/android/issues/detail?id=77830

Change-Id: Idbf68c417d95f7ce2f81abfb0e24ffe9463ac350