Handling for possible in-use files in SdkQuickfixWizard
The most likely case for us to encounter issues with locked
in-use files is on windows when installing an update. In this
case, we now allow the user to only attempt to install the
new packages, and then launch the standalone SDK manager and
exit studio to install the rest.
This is a stopgap until we have proper updater support for
SDK components.

Change-Id: If07b9a3ffaeda011aa217f16247e3c7f716ce65b