Recipe files can now be parsed in dry-run mode
Before, recipe files, when parsed, would automatically execute their
instructions. However, in a following CL, we'll need to read the
contents of a recipe file before deciding whether we want to execute
its instructions or not.

This is accomplished with the concept of a dry run recipe. If in dry
run mode, a recipe will parse but not execute.

Change-Id: Ied2d1fd5cd48faed6a9d391cfb18b1898b781881