Fixed layout of "SDK changed" dialog.
This CL replaces the current dialog that notifies
the user that the IDE and project SDKs are different,
and instead, uses the standard "Yes/No" dialog.

This new dialog does not have the layout issue as
the original one did, and when used on Mac, it shows
like the rest of dialogs: as a "Mac sheet".

This CL adds UI tests for SDK switching.

Fixes
https://code.google.com/p/android/issues/detail?id=80827

Change-Id: Ia8d5cefcf6c20a512cb4e0bc10b7fe3ff0ce15c5