Fix stale jar ClassLoader bug.
When opening a new project, the list of dependencies isn't up to date.
This results in creation of ClassLoader which cannot resolve
dependencies found later in the sync. Fix the issue by getting the list
of jar dependencies and comparing it with the list of dependencies in
the current jar ClassLoader.

Also fix Exception handling in RenderLog. When intellij's hasn't
finished indexing, the rendering fails. However the error reporting
also fails with IndexNotReadyException when it tries to create a
smarter error message. The result was a failed rendering with no
explanation.

Change-Id: I70b1e05684ba2cf9bfc9e7114637249a6e510fb6