Fixed editor notifications when sync is needed.
The issue was that sync notifications were disabled
in GradleSyncState and when the user types in a gradle
file, the notification that indicates a sync is
needed was not showing up.

This CL fixes this issue by properly enabling
notifications after a sync. It also removes code that
is no longer used (e.g. Ide2GradleProjectSync).

Fixes issue
https://code.google.com/p/android/issues/detail?id=171370

Change-Id: I4957cbbf11686b4b9da0f1b987591eea3a4503a6