property sheet: improve grouping and sorting
This CL implements a first cut at grouping attributes together
and displaying them in a certain order.

Attributes are first grouped into groups based on a few different
criteria such as the parent stylable that declares the attribute,
and for certain attributes coming from the base View class, by their
name such as padding/margin or layout parameters.

The grouped items are then sorted based on a predefined ordering
for known types of attributes.

Both the grouping and sorting needs to be improved further based
on criteria from UX and including per View meta data.

Change-Id: Ie8a9c23f18618a7a255780b6cade6db4461fe79a