Nele: Drop the new component on the correct screen view.
There were a bug: if you drag a new component in from the left
(i.e. over the blueprint) and drop it on the normal screen view,
the coordinates would be computed as if it was dropped on the
blueprint view.

Change-Id: Ib5254c534e7457541211db089a543a163e0fcc27
(cherry picked from commit acc7d0f37dc3ee2e36344ecb3395dc1ed7c2409b)