Do not access attributes directly from the property sheet.
The XML tag and/or attribute information could be invalid at a later
point in time if the user has performed some edit. Rather than directly
accessing the tag/attribute and checking for its validity, we query
the component for this information. It is now the component's
responsibility to check for validity. The component should eventually be
updated to maintain a copy of the PSI data structure so that it can
always return the result from a snapshot in time.

Change-Id: Id002993503227412dc32968e87e84c25c5a9352c
(cherry picked from commit b2494311070662d23a16a8309c2cad5cfde88e86)