Custom renderer for android resource ids
This CL adds a custom node renderer for integer values that have
one of the following android annotations: @IntDef, @ColorInt or
one of the many @*Res.

Currently, this can only detect fields and method parameters that
have this annotation (i.e no local variables).

In the case of @*Res annotation, we only show the resource names
corresponding to the framework resources for now.

Change-Id: Ife6169e112e60590afc608a450861a224a68f1cd