Last custom source couldn't be removed
Previously it was impossible to remove all custom SDK sources:
the last one would reappear after being removed, since a null or empty
list was taken as a sign to reload the custom sites.
This change sets the list to a special marker when all sites have been
removed to indicate that it shouldn't be reloaded.

Change-Id: Iba4427000e49515f16aaaacbee9645886c9681c3